requirements.txt
anyio==3.7.1
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
fastapi==0.103.2
h11==0.14.0
idna==3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.33
pip==23.3
protobuf==3.20.3
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pyjwt==2.8.0  
sniffio==1.3.0
starlette==0.27.0
typing==3.7.4.3
typing-extensions==4.7.1
uvicorn==0.15.0  
jwt == 1.3.1
pyodbc==5.1.0


